+++ tests/cd.rs
@@ -17,6 +17,7 @@ use std::rc::Rc;
 #[macro_use]
 mod support;
 pub use self::support::*;
+pub use self::support::spawn::builtin::cd;
 
 struct CdResult {
     initial_cwd: PathBuf,
@@ -45,7 +46,7 @@ fn run_cd<F>(cd_args: &[&str], env_setup: F) -> CdResult
     let read_to_end_out = tokio_io::io::read_to_end(read_to_end_out, Vec::new());
     let read_to_end_err = tokio_io::io::read_to_end(read_to_end_err, Vec::new());
 
-    let mut cd = builtin::cd(cd_args.iter().map(|&s| s.to_owned()))
+    let mut cd = cd(cd_args.iter().map(|&s| s.to_owned()))
         .spawn(&env);
 
     let env = RefCell::new(env);
@@ -84,7 +85,7 @@ fn successful_if_no_stdout() {
     let (mut lp, mut env) = new_env_with_no_fds();
 
     let args: Vec<Rc<String>> = vec!(input.to_string_lossy().into_owned().into());
-    let mut cd = builtin::cd(args)
+    let mut cd = cd(args)
         .spawn(&env);
 
     let exit = lp.run(poll_fn(|| cd.poll(&mut env)).flatten());
@@ -289,7 +290,7 @@ fn dash_unset_old_pwd_is_error() {
 #[should_panic]
 fn polling_canceled_pwd_panics() {
     let (_, mut env) = new_env_with_no_fds();
-    let mut cd = builtin::cd(Vec::<Rc<String>>::new())
+    let mut cd = cd(Vec::<Rc<String>>::new())
         .spawn(&env);
 
     cd.cancel(&mut env);