+++ tests/false_cmd.rs
@@ -1,11 +1,12 @@
 mod support;
 pub use self::support::*;
+pub use self::support::spawn::builtin::false_cmd;
 
 #[test]
 fn false_smoke() {
     let (mut lp, mut env) = new_env();
 
-    let exit = lp.run(builtin::false_cmd()
+    let exit = lp.run(false_cmd()
         .spawn(&env)
         .pin_env(&mut env)
         .flatten()