+++ tests/pwd.rs
@@ -14,6 +14,7 @@ use std::rc::Rc;
 #[macro_use]
 mod support;
 pub use self::support::*;
+pub use self::support::spawn::builtin::pwd;
 
 #[derive(Debug, Clone)]
 struct DummyWorkingDirEnv(PathBuf);
@@ -65,7 +66,7 @@ fn run_pwd(use_dots: bool, pwd_args: &[&str], physical_result: bool) {
         .change_working_dir_env(DummyWorkingDirEnv(cur_dir))
     );
 
-    let pwd = builtin::pwd(pwd_args.iter().map(|&s| s.to_owned()));
+    let pwd = pwd(pwd_args.iter().map(|&s| s.to_owned()));
 
     let pipe = env.open_pipe().expect("pipe failed");
     env.set_file_desc(conch_runtime::STDOUT_FILENO, pipe.writer, Permissions::Write);
@@ -131,7 +132,7 @@ fn last_specified_flag_wins() {
 #[test]
 fn successful_if_no_stdout() {
     let (mut lp, env) = new_env_with_no_fds();
-    let pwd = builtin::pwd(Vec::<Rc<String>>::new());
+    let pwd = pwd(Vec::<Rc<String>>::new());
     let exit = lp.run(pwd.spawn(&env).pin_env(env).flatten());
     assert_eq!(exit, Ok(EXIT_SUCCESS));
 }
@@ -140,7 +141,7 @@ fn successful_if_no_stdout() {
 #[should_panic]
 fn polling_canceled_pwd_panics() {
     let (_, mut env) = new_env_with_no_fds();
-    let mut pwd = builtin::pwd(Vec::<Rc<String>>::new())
+    let mut pwd = pwd(Vec::<Rc<String>>::new())
         .spawn(&env);
 
     pwd.cancel(&mut env);