+++ tests/shift.rs
@@ -5,6 +5,7 @@ use std::rc::Rc;
 
 mod support;
 pub use self::support::*;
+pub use self::support::spawn::builtin::shift;
 
 fn run_shift(
     env_args_starting: &[&str],
@@ -20,7 +21,7 @@ fn run_shift(
         .collect::<Vec<_>>()
     ));
 
-    let shift = builtin::shift(shift_args.iter().map(|&s| s.to_owned()));
+    let shift = shift(shift_args.iter().map(|&s| s.to_owned()));
 
     let mut shift = shift.spawn(&env);
     let exit = lp.run(poll_fn(|| shift.poll(&mut env)).flatten())
@@ -75,7 +76,7 @@ fn shift_multiple_arg_does_nothing_and_exit_with_error() {
 #[should_panic]
 fn polling_canceled_shift_panics() {
     let (_, mut env) = new_env_with_no_fds();
-    let mut shift = builtin::shift(Vec::<Rc<String>>::new())
+    let mut shift = shift(Vec::<Rc<String>>::new())
         .spawn(&env);
 
     shift.cancel(&mut env);